Initial dvach (2ch.hk) support (can load catalogs/threads/some other minor stuff)

+++ app/src/main/java/com/github/k1rakishou/kurobaexlite/model/data/IPostImage.kt
@@ -16,7 +16,7 @@ interface IPostImage {
   val ext: String
   val width: Int
   val height: Int
-  val fileSize: Int
+  val fileSize: Long
   val thumbnailSpoiler: ThumbnailSpoiler?
 
   val ownerPostDescriptor: PostDescriptor
