+++ app/src/main/java/com/github/k1rakishou/kurobaexlite/model/data/local/PostImageData.kt
@@ -15,7 +15,7 @@ data class PostImageData(
   override val ext: String,
   override val width: Int,
   override val height: Int,
-  override val fileSize: Int,
+  override val fileSize: Long,
   override val thumbnailSpoiler: ThumbnailSpoiler?,
   override val ownerPostDescriptor: PostDescriptor
 ) : IPostImage {
