+++ app/src/main/java/com/github/k1rakishou/kurobaexlite/model/data/ui/post/PostCellImageData.kt
@@ -16,7 +16,7 @@ data class PostCellImageData(
   override val ext: String,
   override val width: Int,
   override val height: Int,
-  override val fileSize: Int,
+  override val fileSize: Long,
   override val thumbnailSpoiler: ThumbnailSpoiler?,
   override val ownerPostDescriptor: PostDescriptor
 ) : IPostImage {
